Consolidate Japanese sentence tokenizer and regular sentence tokenizer into one
