+++ packages/yoastseo/src/languageProcessing/helpers/sentence/SentenceTokenizer.js
@@ -12,8 +12,18 @@ const fullStop = ".";
  * \u2026 - ellipsis.
  * \u06D4 - Urdu full stop.
  * \u061f - Arabic question mark.
+ * \u3002 - Japanese ideographic full stop.
+ * \uFF61 - Japanese half-width ideographic full stop.
+ * \uFF01 - Japanese full-width exclamation mark.
+ * \u203C - Japanese double exclamation mark.
+ * \uFF1F - Japanese fullwidth question mark.
+ * \u2047 - Japanese double question mark.
+ * \u2049 - Japanese exclamation question mark.
+ * \u2048 - Japanese question exclamation mark.
+ * \u2026 - Japanese horizontal ellipsis.
+ * \u2025 - Japanese two dot leader.
  */
-const sentenceDelimiters = "?!;\u2026\u06d4\u061f";
+const sentenceDelimiters = "?!;\u2026\u06d4\u061f\u3002\uFF61\uFF01\u203C\uFF1F\u2047\u2049\u2048\u2049\u2026\u2025";
 
 const fullStopRegex = new RegExp( "^[" + fullStop + "]$" );
 const sentenceDelimiterRegex = new RegExp( "^[" + sentenceDelimiters + "]$" );
@@ -38,7 +48,18 @@ export default class SentenceTokenizer {
 	 * @returns {boolean} Whether or not the character is a capital letter.
 	 */
 	isNumber( character ) {
-		return ! isNaN( parseInt( character, 10 ) );
+		const japaneseNumbers = [
+			// Full-width.
+			/^[\uFF10-\uFF19]+$/i,
+			// Circled digit, parenthesized digit, and digit with full stop.
+			/^[\u2460-\u249B]+$/i,
+			// Parenthesized ideograph.
+			/^[\u3220-\u3229]+$/i,
+			// Circled ideograph.
+			/^[\u3280-\u3289]+$/i,
+		];
+
+		return ( ! isNaN( parseInt( character, 10 ) ) || japaneseNumbers.some( numberRange => numberRange.test( character ) ) );
 	}
 
 	/**
@@ -62,6 +83,18 @@ export default class SentenceTokenizer {
 
 		return "'" === character ||
 			"\"" === character;
+	}
+
+	/**
+	 * Returns whether or not a given character is a Japanese quotation mark.
+	 *
+	 * @param {string} character The character to check.
+	 * @returns {boolean} Whether or not the given character is a Japanese quotation mark.
+	 */
+	isJapaneseQuotation( character ) {
+		const openingQuotationMark = /^[\u300C\u300E\u3008\u3014\u3010\uFF5B\uFF3B]+$/i;
+
+		return openingQuotationMark.test( character );
 	}
 
 	/**
@@ -157,6 +190,31 @@ export default class SentenceTokenizer {
 	}
 
 	/**
+	 * Checks whether a character is from Japanese language that could be sentence beginning.
+	 *
+	 * @param {string} letter The letter to check.
+	 *
+	 * @returns {boolean} Whether the letter is from Japanese language that could be sentence beginning.
+	 */
+	isJapaneseSentenceBeginning( letter ) {
+		const japaneseLetterRanges = [
+			// Hiragana.
+			/^[\u3040-\u3096]+$/i,
+			// Katakana full-width.
+			/^[\u30A1-\u30FA]+$/i,
+			/^[\u31F0-\u31FF]+$/i,
+			// Katakana half-width.
+			/^[\uFF66-\uFF9D]+$/i,
+			// Kanji.
+			/^[\u4E00-\u9FFC]+$/i,
+		];
+
+		return (
+			japaneseLetterRanges.some( letterRange => letterRange.test( letter ) )
+		);
+	}
+
+	/**
 	 * Checks if the sentenceBeginning beginning is a valid beginning.
 	 *
 	 * @param {string} sentenceBeginning The beginning of the sentence to validate.
@@ -168,7 +226,9 @@ export default class SentenceTokenizer {
 				this.isNumber( sentenceBeginning ) ||
 				this.isQuotation( sentenceBeginning ) ||
 				this.isPunctuation( sentenceBeginning ) ||
-				this.isSmallerThanSign( sentenceBeginning ) );
+				this.isSmallerThanSign( sentenceBeginning ) ||
+				this.isJapaneseSentenceBeginning( sentenceBeginning ) ||
+				this.isJapaneseQuotation( sentenceBeginning ) );
 	}
 
 	/**
